+++ js/effects.js
@@ -1,0 +1,1 @@
+//Наложение эффектов на изображение.
