+++ js/filter.js
@@ -1,0 +1,1 @@
+// Фильтрация изображений от других пользователей.
