+++ js/form-validation.js
@@ -1,0 +1,1 @@
+// Валидация формы.
