+++ js/scale.js
@@ -1,0 +1,1 @@
+// Изменение масштаба изображения.
