+++ js/upload_file.js
@@ -1,0 +1,1 @@
+// Загрузка изображений.
